Added checkDiff

+++ src/Database/difficulty.h
@@ -2,6 +2,6 @@
 #define DIFF_H
 #include <stdint.h>
 
-uint64_t diff(uint64_t* difficulties, uint64_t* timestamps, uint8_t inDiffCount);
+uint64_t diff(uint64_t* difficulties, uint32_t* timestamps, uint8_t inDiffCount);
 
 #endif
